+++ app/src/main/res/values/strings.xml
@@ -2007,4 +2007,6 @@
     <string name="recommended_reading_list_notification_subtitle">Tap to see your latest recommendations.</string>
     <string name="recommended_reading_list_settings_notification_subtitle_enable">When your Discover is updated, Wikipedia will send you a notification.</string>
     <string name="recommended_reading_list_settings_toggle_enable_message">Discover is on. Wikipedia will recommend you articles to read.</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
